refactor: support message length of uint32_t

+++ ST25DV64KC.h
@@ -130,5 +130,5 @@ void initializeI2C();
 
 //Send a message given an array and its length
 //Set prependAddressFlag if you wish to have write address automatically calculated and injected
-I2C_Status sendMessage(uint8_t * message, uint8_t * messageLength, uint8_t prependAddressFlag);
+I2C_Status sendMessage(uint8_t * message, uint32_t * messageLength, uint8_t prependAddressFlag);
 #endif
